refactor(pattern): remove legacy pattern engine
