+++ .trellis/tasks/06-12-consolidate-pattern-runtime-engine/task.json
@@ -1,0 +1,26 @@
+{
+  "id": "consolidate-pattern-runtime-engine",
+  "name": "consolidate-pattern-runtime-engine",
+  "title": "Consolidate Pattern Runtime Engine",
+  "description": "",
+  "status": "in_progress",
+  "dev_type": null,
+  "scope": null,
+  "package": null,
+  "priority": "P2",
+  "creator": "cxzzzz",
+  "assignee": "cxzzzz",
+  "createdAt": "2026-06-12",
+  "completedAt": null,
+  "branch": null,
+  "base_branch": "feature/programmable_pattern",
+  "worktree_path": null,
+  "commit": null,
+  "pr_url": null,
+  "subtasks": [],
+  "children": [],
+  "parent": null,
+  "relatedFiles": [],
+  "notes": "",
+  "meta": {}
+}
