+++ src/wavekit/pattern/__init__.py
@@ -1,7 +1,7 @@
 """Protocol pattern matching for waveform analysis."""
 
 from .dsl import Pattern as Pattern
-from .engine import PatternError as PatternError
+from .errors import PatternError as PatternError
 from .result import MatchResult as MatchResult
 from .result import MatchStatus as MatchStatus
 from .steps import Channel as Channel
