+++ src/wavekit/pattern/errors.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+class PatternError(Exception):
+    """Raised for pattern definition or runtime errors."""
